Module Gemini

A strongly typed gemini trading api written in pure OCaml with both websocket and rest endpoints supported.

Submodules of the V1 module are either REST operations which can be invoked via a well typed invocation of Operation.post, or they are web socket interfaces such as Market_data and Order_events. Use the client function to get a typed response pipe over the socket of these services.

All service invocations require a Cfg module which is usually provided from the command line or environment variables for api host and secret information.

module Auth : sig ... end

module Cfg : sig ... end

module Nonce : sig ... end

module Rest : sig ... end

REST api support for the Gemini trading exchange. These endpoints are used to manage orders and check balances.

module Result : sig ... end

module Inf_pipe : sig ... end

module Poly_ok : sig ... end

module V1 : sig ... end

Version v1 of the Gemini REST and web socket apis.