Module Cfg.Production

Creates a configuration module for the production environment. Will query the unix environment for the necessary parameters.

It is recommended to generate this module exactly once in your program.

Parameters

Signature

include S

val version : string

val api_host : string

val api_key : string

val api_secret : string