Module type Cfg.S

Gemini configuration. These values are the only necessary parameters and secrets to connect to an endpoint.

For each gemini environment a different configuration module should be instantiated.

val version : string

val api_host : string

val api_key : string

val api_secret : string