Module Gemini.Rest

REST api support for the Gemini trading exchange. These endpoints are used to manage orders and check balances.

module Error : sig ... end

Represents all error conditions possible by the REST services. In general the REST services don't throw exceptions- instead they report an error variant with some provisional failure information.

module Request = Nonce.Request

module Operation : sig ... end

Operations denote a single REST operation endpoint.

module Response : sig ... end

Support for parsing responses from a Gemini api REST endpoint.

module Post (Operation : Operation.S) : sig ... end

Creates a REST endpoint using the http post method with operation specified by provided module Operation

module Make (Operation : Operation.S) : sig ... end

Creates a REST endpoint using the http post method with operation specified by provided module Operation. Also produces a command line interface hook.

module Make_no_arg (Operation : Operation.S_NO_ARG) : sig ... end

Creates a REST endpoint using the http post method with no argument operation specified by provided module Operation. Also produces a command line interface hook.