Parameter Make_no_arg.Operation

include Operation.S with type request = unit

val name : string

The human readable name of this operation.

val path : string list

The uri path of the REST endpoint.

type request = unit

The type of the request payload for this REST endpoint.

val request_to_yojson : request -> Yojson.Safe.t

val sexp_of_request : request -> Sexplib0.Sexp.t

val request_of_sexp : Sexplib0.Sexp.t -> request

type response

The type of the response payload for this REST endpoint.

val response_of_yojson : Yojson.Safe.t -> response Ppx_deriving_yojson_runtime.error_or

val sexp_of_response : response -> Sexplib0.Sexp.t

val response_of_sexp : Sexplib0.Sexp.t -> response