Module Gemini.V1

Version v1 of the Gemini REST and web socket apis.

val path : string list

module Heartbeat : sig ... end

Heartbeat REST api operation.

module Timestamp : sig ... end

module Currency : sig ... end

module Symbol : sig ... end

module Exchange : sig ... end

module Side : sig ... end

module Order_type : sig ... end

module Order_execution_option : sig ... end

Represents different execution rules for an order.

module Order : sig ... end

Represents an order on the Gemini trading exchange over the REST api.

module Orders : sig ... end

Gets the status of all open orders on Gemini trading exchange over the REST api.

module Mytrades : sig ... end

Gets all trades executed by this api user on Gemini trading exchange over the REST api.

module Tradevolume : sig ... end

Gets all trade volume executed by on the Gemini trading exchange over the REST api.

module Balances : sig ... end

Gets all balances for this api user on the Gemini trading exchange over the REST api.

module Notional_volume : sig ... end

module Market_data : sig ... end

val command : Async.Command.t