Module V1.Balances

Gets all balances for this api user on the Gemini trading exchange over the REST api.

type request = unit

Balance queries have no request parameters.

val request_of_yojson : Yojson.Safe.t -> request Ppx_deriving_yojson_runtime.error_or

type balance = {

currency : Currency.Enum_or_string.t;

amount : string;

available : string;

available_for_withdrawal : string;

type_ : string;

}

The type of a balance for one specific currency.

val balance_to_yojson : balance -> Yojson.Safe.t

val balance_of_yojson : Yojson.Safe.t -> balance Ppx_deriving_yojson_runtime.error_or

val sexp_of_balance : balance -> Sexplib0.Sexp.t

val balance_of_sexp : Sexplib0.Sexp.t -> balance

type response = balance list

include Rest.Operation.S with type request := request with type response := response

val name : string

The human readable name of this operation.

val path : string list

The uri path of the REST endpoint.

val request_to_yojson : request -> Yojson.Safe.t

val sexp_of_request : request -> Sexplib0.Sexp.t

val request_of_sexp : Sexplib0.Sexp.t -> request

val response_of_yojson : Yojson.Safe.t -> response Ppx_deriving_yojson_runtime.error_or

val sexp_of_response : response -> Sexplib0.Sexp.t

val response_of_sexp : Sexplib0.Sexp.t -> response

val post : (module Cfg.S) -> Nonce.reader -> request -> [ Rest.Error.post | `Ok of response ] Async.Deferred.t

val command : string * Async.Command.t