Module Side.Bid_ask

Represents a bid or ask side.

type t = [

| `Ask

| `Bid

]

val __t_of_sexp__ : Sexplib0.Sexp.t -> t

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

include Ppx_enumerate_lib.Enumerable.S with type t := t

val all : t list

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

module Enum_or_string : sig ... end

Wraps a string value promoting it to type Enum enum if it is parsable as such, `String string otherwise.

val dict : (string * t) list

val of_string_opt : string -> t option

val error_message : string -> string

include Csvfields.Csv.Stringable with type t := t

val to_string : t -> string

val of_string : string -> t

val is_csv_atom : bool

val rev_csv_header' : string list -> 'a -> 'b -> string list

val rev_csv_header_spec' : Csvfields.Csv.Spec.t list -> 'a -> 'b -> Csvfields.Csv.Spec.t list

val t_of_row' : 'a -> string list -> (unit -> t) * string list

val write_row_of_t' : is_first:bool -> is_last:bool -> writer:(string -> unit) -> 'a -> 'b -> t -> unit

val csv_header : string list

val csv_header_spec : Csvfields.Csv.Spec.t list

val t_of_row : string list -> t

val row_of_t : t -> string list

val csv_load : ?separator:char -> string -> t list

val csv_load_in : ?separator:char -> Core.In_channel.t -> t list

val csv_save_fn : ?separator:char -> (string -> unit) -> t list -> unit

val csv_save_out : ?separator:char -> Core.Out_channel.t -> t list -> unit

val csv_save : ?separator:char -> string -> t list -> unit