Module V1.Mytrades

Gets all trades executed by this api user on Gemini trading exchange over the REST api.

type trade = {

price : string;

amount : string;

timestamp : Timestamp.Sec.t;

timestampms : Timestamp.Ms.t;

type_ : Side.t;

aggressor : bool;

fee_currency : Currency.Enum_or_string.t;

fee_amount : string;

tid : int64;

order_id : int64;

client_order_id : string option;

is_auction_fill : bool;

is_clearing_fill : bool;

symbol : Symbol.Enum_or_string.t;

exchange : Exchange.t;

}

Represents one instance of a trade exchanged on Gemini.

val trade_to_yojson : trade -> Yojson.Safe.t

val trade_of_yojson : Yojson.Safe.t -> trade Ppx_deriving_yojson_runtime.error_or

val sexp_of_trade : trade -> Sexplib0.Sexp.t

val trade_of_sexp : Sexplib0.Sexp.t -> trade

type request = {

symbol : Symbol.t;

limit_trades : int option;

timestamp : Timestamp.Sec.t option;

}

Trade request parameters.

val request_of_yojson : Yojson.Safe.t -> request Ppx_deriving_yojson_runtime.error_or

type response = trade list

Mytrades repsonse type- the type of a list trades.

include Rest.Operation.S with type request := request with type response := response

val name : string

The human readable name of this operation.

val path : string list

The uri path of the REST endpoint.

val request_to_yojson : request -> Yojson.Safe.t

val sexp_of_request : request -> Sexplib0.Sexp.t

val request_of_sexp : Sexplib0.Sexp.t -> request

val response_of_yojson : Yojson.Safe.t -> response Ppx_deriving_yojson_runtime.error_or

val sexp_of_response : response -> Sexplib0.Sexp.t

val response_of_sexp : Sexplib0.Sexp.t -> response

val post : (module Cfg.S) -> Nonce.reader -> request -> [ Rest.Error.post | `Ok of response ] Async.Deferred.t

val command : string * Async.Command.t