Module V1.Order

Represents an order on the Gemini trading exchange over the REST api.

val name : string

val path : string list

module Status : sig ... end

Represents the status of an order on the Gemini trading exchange over the REST api.

module New : sig ... end

Represents a new order request on the Gemini trading exchange over the REST api.

module Cancel : sig ... end

Represents order cancellation features on the Gemini trading exchange over the REST api.

val command : string * Async.Command.t