Module V1.Tradevolume

Gets all trade volume executed by on the Gemini trading exchange over the REST api.

type volume = {

account_id : int64 option;

symbol : Symbol.Enum_or_string.t;

base_currency : Currency.Enum_or_string.t;

notional_currency : Currency.Enum_or_string.t;

data_date : string;

total_volume_base : float;

maker_buy_sell_ratio : float;

buy_maker_base : float;

buy_maker_notional : float;

buy_maker_count : int64;

sell_maker_base : float;

sell_maker_notional : float;

sell_maker_count : int64;

buy_taker_base : float;

buy_taker_notional : float;

buy_taker_count : int64;

sell_taker_base : float;

sell_taker_notional : float;

sell_taker_count : int64;

}

The type of a trade volume entity for one particular symbol on the Gemini trading exchange.

val volume_to_yojson : volume -> Yojson.Safe.t

val volume_of_yojson : Yojson.Safe.t -> volume Ppx_deriving_yojson_runtime.error_or

val sexp_of_volume : volume -> Sexplib0.Sexp.t

val volume_of_sexp : Sexplib0.Sexp.t -> volume

type request = unit

val request_of_yojson : Yojson.Safe.t -> request Ppx_deriving_yojson_runtime.error_or

type response = volume list list

The type of a trade volume response, which is a list of list of volumes grouped by exchange symbol.

include Rest.Operation.S with type request := request with type response := response

val name : string

The human readable name of this operation.

val path : string list

The uri path of the REST endpoint.

val request_to_yojson : request -> Yojson.Safe.t

val sexp_of_request : request -> Sexplib0.Sexp.t

val request_of_sexp : Sexplib0.Sexp.t -> request

val response_of_yojson : Yojson.Safe.t -> response Ppx_deriving_yojson_runtime.error_or

val sexp_of_response : response -> Sexplib0.Sexp.t

val response_of_sexp : Sexplib0.Sexp.t -> response

val post : (module Cfg.S) -> Nonce.reader -> request -> [ Rest.Error.post | `Ok of response ] Async.Deferred.t

val command : string * Async.Command.t